requirements.txt
libvirt-python==8.1.0
